aurora/server/requirements.txt
asgiref==3.11.0
backtesting==0.6.5
bokeh==3.8.1
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
massive==2.0.2
narwhals==2.13.0
numpy==2.3.5
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg==3.2.13
psycopg-binary==3.2.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
xyzservices==2025.11.0
